+++ src/datasources/data-frame/types.ts
@@ -374,6 +374,11 @@ export interface CustomPropertyOptions {
   columnCustomPropertyOptions: Option[];
 }
 
+export interface PropertySelections {
+  customProperties: string[];
+  standardProperties: DataTableProperties[];
+}
+
 export interface Option {
   label: string;
   value: string;
@@ -410,3 +415,7 @@ export interface TableColumnsData {
 export function isSystemLinkError(error: any): error is SystemLinkError {
   return Boolean(error?.error?.code) && Boolean(error?.error?.name);
 }
+
+export const ALL_STANDARD_PROPERTIES = new Set<string>(
+  Object.values(DataTableProperties)
+);